# design/openapi.yaml
...
openapi: 3.0.3
info:
  title: DriveDeal — Reverse Car Marketplace API
  version: 1.0.0
  description: |-
    Backend API for DriveDeal: a reverse car marketplace where buyers post the
    car they want and dealers compete on price for it.

    Handles:
    - Buyer flows: post a car request, watch the live quote leaderboard, accept a
      quote, track the deal, chat with the dealer.
    - Dealer flows: account with manual approval, geo-filtered request feed, quote
      submission and revision, deal status.
    - Support flows: ticket queue, dealer/agent verification approval, roster.
    - AI flows (buyer only): a streaming advisory agent and a compare agent,
      grounded in the database plus a Crawl4AI web search.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the /api/v1/ws WebSocket.
    - **Success responses are envelope-free.** The payload is the resource itself.
      There is no {"success": true, "data": ...} wrapper — the frontend's mock
      client already returns bare objects and that shape is the contract
      (solution-frontend.md §2.1, solution-backend.md §0.2).
    - **Every non-2xx JSON response uses the error envelope**, and nothing else:
        {"error": {"code", "message", "details", "request_id"}}
      `code` is a stable machine-readable string the client branches on.
      `message` is for humans and may be reworded; never parse it.
...
  contact:
...
servers:
...
security:
- bearer_auth: []
paths:
...
  /support/tickets:
    post:
      summary: Raise a support ticket
      description: |-
        `category` is derived from the caller's role, never accepted from the
        body — a dealer cannot file into the customer stream. `ticket_id` is
        generated server-side: `TIC-<epoch>` for customers, `DS<epoch>` for
        dealers.
      operationId: createTicket
      tags:
      - Support
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
            examples:
              slow_site:
                value:
                  issue_summary: The web page is very slow
                  issue_description: Pages take 8+ seconds since yesterday evening.
                  priority: medium
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/unauthorized'
        '422':
          $ref: '#/components/responses/validation_error'
    get:
      summary: List my tickets
      description: |-
        Callers see only their own tickets (`caller_id = me`). Cursor
        pagination.
      operationId: listMyTickets
      tags:
      - Support
      parameters:
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/TicketStatus'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicketList'
        '401':
          $ref: '#/components/responses/unauthorized'
  /support/tickets/{id}:
    get:
      summary: Get a ticket
      description: |-
        The caller who raised it, or a support/admin agent. A dealer's ticket is
        scoped to `category = 'dealer'`, so the two streams stay separated.
      operationId: getTicket
      tags:
      - Support
      parameters:
      - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
  /support/queue/tickets:
    get:
      summary: Support agent work queue
      description: |-
        Support/admin only. The agent's home view: all tickets, ordered by
        priority then age, with the SLA clock. `category` defaults to
        `customer` so an agent working the customer queue does not see dealer
        tickets unless they ask.
      operationId: getSupportQueue
      tags:
      - Support
      parameters:
      - name: category
        in: query
        schema:
          type: string
          enum:
          - customer
          - dealer
          - internal
          default: customer
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/TicketStatus'
      - name: assigned_to
        in: query
        description: Restrict to one agent
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Queue page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicketList'
              examples:
                queue:
                  value:
                    items:
                    - $ref: '#/components/examples/TicketNew'
                    - $ref: '#/components/examples/TicketInProgress'
                    next_cursor: null
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /support/queue/tickets/{id}:
    patch:
      summary: Update a ticket from the queue
      description: |-
        Support/admin only. Appends to `notes` (a jsonb timeline, never
        overwritten) and may set `rca`. Resolving a ticket sets
        `resolved_at`; closing it sets `closed_at`.
      operationId: updateQueuedTicket
      tags:
      - Support
      parameters:
      - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueTicketUpdateRequest'
            examples:
              agent_note:
                value:
                  status: in-progress
                  assigned_to: 4772bb1a-ee75-4f45-b8e6-53d70f6e42b1
                  note:
                    text: Reproduced on staging — DB query missing the brand index.
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
  /verifications:
    get:
      summary: List verification requests
      description: |-
        Support/admin only. Defaults to `status = 'pending'`, because the
        actionable work is always what is waiting on a human.
      operationId: listVerifications
      tags:
      - Verifications
      parameters:
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/VerificationStatus'
      - name: category
        in: query
        schema:
          type: string
          enum:
          - dealer
          - agent
          - customer
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of verifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationList'
              examples:
                pending:
                  value:
                    items:
                    - $ref: '#/components/examples/VerificationPending'
                    next_cursor: null
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /verifications/{id}/approve:
    post:
      summary: Approve a verification
      description: |-
        The single gate that makes an account usable. Approving a **dealer**
        verification sets `users.is_active = true` and `verified_at` in the same
        transaction, so the dealer's next login succeeds and their quotes can
        start flowing. `email_sent` flips to true once the notification is out.

        **Support agents may approve dealer verifications only. Approving an
        `agent` verification requires the `admin` role** — 403
        INSUFFICIENT_ROLE for a support agent. Agents cannot approve peers.
      operationId: approveVerification
      tags:
      - Verifications
      parameters:
      - $ref: '#/components/parameters/VerificationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationDecisionRequest'
            examples:
              approve_dealer:
                value:
                  note: License TX030202 verified against the state registry.
      responses:
        '200':
          description: Verification approved, account activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verification'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /verifications/{id}/deny:
    post:
      summary: Deny a verification
      description: |-
        Support/admin only, same role rule as approve. The account stays
        inactive; the applicant is emailed the reason and may apply again.
      operationId: denyVerification
      tags:
      - Verifications
      parameters:
      - $ref: '#/components/parameters/VerificationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationDecisionRequest'
            examples:
              deny:
                value:
                  note: Dealership licence number does not match the state registry.
      responses:
        '200':
          description: Verification denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verification'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /members:
    get:
      summary: List members
      description: |-
        Support/admin only. Every profile, any role, with account state. This
        is the roster screen; buyer and dealer detail is not served from the
        public /profiles/{id}.
      operationId: listMembers
      tags:
      - Members
      parameters:
      - name: role
        in: query
        schema:
          $ref: '#/components/schemas/Role'
      - name: is_active
        in: query
        schema:
          type: boolean
      - name: q
        in: query
        description: Case-insensitive match on full_name, email or phone
        schema:
          type: string
          maxLength: 100
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /members/{id}/suspend:
    post:
      summary: Suspend a member
      description: |-
        Support/admin only. Sets `users.is_active = false` and revokes every
        refresh token family, so access dies within one access-token lifetime.
        Their rows are never deleted — the deal history has to stay intact.
      operationId: suspendMember
      tags:
      - Members
      parameters:
      - $ref: '#/components/parameters/ProfileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - reason
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 1000
      responses:
        '200':
          description: Member suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
...
components:
  parameters:
    ProfileId:
      name: id
      in: path
      required: true
      description: Profile UUID
      schema:
        type: string
        format: uuid
...
    TicketId:
      name: id
      in: path
      required: true
      description: Support ticket UUID
      schema:
        type: string
        format: uuid
    VerificationId:
      name: id
      in: path
      required: true
      description: Verification UUID
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from the previous page
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        RS256 access token from /auth/login or /auth/refresh.

        Subject = `profiles.id` (NOT the `users.id`, which is a different
        UUID and does not appear anywhere in the request layer). Roles are
        read from the `profiles` row on every request, so a role change or a
        suspension takes effect on the next call rather than at token expiry.

        For the WebSocket, use a one-shot ticket from /chats/ws-ticket
        instead — a browser cannot set this header on the handshake.
  responses:
    unauthorized:
      description: |-
        Missing, malformed, expired, or revoked token. `code`
        distinguishes: `UNAUTHORIZED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Your session has expired. Sign in again.
                  details: null
                  request_id: 3f2a1b0c-9d8e-4f7a-8b6c-5d4e3f2a1b0c
    forbidden:
      description: |-
        Authenticated, but not allowed. `code` distinguishes:
        `FORBIDDEN_ROLE` (wrong role for this surface), `NOT_RESOURCE_OWNER`
        (someone else's resource), `INSUFFICIENT_ROLE` (a support agent trying
        to approve a peer agent).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            wrong_role:
              value:
                error:
                  code: FORBIDDEN_ROLE
                  message: The AI advisor is only available to buyers.
                  details: null
                  request_id: 7a6b5c4d-3e2f-4a1b-8c9d-0e1f2a3b4c5d
            agent_approving_agent:
              value:
                error:
                  code: INSUFFICIENT_ROLE
                  message: Only an admin can approve a support agent account.
                  details:
                    required_role: admin
                    your_role: support
                  request_id: 1b2c3d4e-5f60-4a1b-8c2d-3e4f5a6b7c8d
...
    not_found:
      description: The resource does not exist, or the caller may not know it exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              value:
                error:
                  code: NOT_FOUND
                  message: That request could not be found.
                  details: null
                  request_id: 4d5e6f70-8a9b-4c1d-9e2f-3a4b5c6d7e8f
    conflict:
      description: |-
        The request is valid but the current state forbids it. `code`
        distinguishes: `CONFLICT` (generic, `details` lists the legal
        transitions), `ALREADY_EXISTS` (duplicate signup, second live quote
        on the same request), `INVALID_TRANSITION` (deal or quote status
        machine rejected the move), `TOKEN_ALREADY_USED` (refresh replay).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            illegal_deal_transition:
              value:
                error:
                  code: INVALID_TRANSITION
                  message: A matched deal cannot move straight to completed.
                  details:
                    current: matched
                    attempted: completed
                    allowed_transitions:
                    - negotiation
                    - cancelled
                  request_id: 5e6f7081-9a2b-4c3d-8e4f-5a6b7c8d9e0f
            duplicate_quote:
              value:
                error:
                  code: ALREADY_EXISTS
                  message: You already have a live quote on this request. Revise it instead.
                  details:
                    existing_quote_id: 724c2d61-b1be-47ba-9458-0f4001204a1b
                  request_id: 6f708192-0a3b-4d5e-9f6a-7b8c9d0e1f2a
    validation_error:
      description: |-
        Body, path or query failed validation. `details` is a
        field → message map so the client can highlight the offending input
        rather than showing a generic banner.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            field_errors:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Some fields need attention.
                  details:
                    email:
                    - That email is already registered.
                    budget_max:
                    - budget_max must be greater than or equal to budget_min.
                  request_id: 0718293a-4b5c-4d6e-8f70-9a1b2c3d4e5f
...
  schemas:
...
    Role:
      type: string
      enum:
      - buyer
      - dealer
      - support
      - admin
      description: |-
        Read from the `profiles` row on every request. Note this is
        `dealer`, **not** `seller` — the schema's `cars.seller_id` column
        name is historical and does not change the role vocabulary.
...
    TicketStatus:
      type: string
      enum:
      - new
      - in-progress
      - on-hold
      - resolved
      - closed
    TicketPriority:
      type: string
      enum:
      - low
      - medium
      - high
      - urgent
    VerificationStatus:
      type: string
      enum:
      - pending
      - approved
      - denied
      - rejected
...
    TicketCreateRequest:
      type: object
      required:
      - issue_summary
      properties:
        issue_summary:
          type: string
          minLength: 1
          maxLength: 200
          example: The web page is very slow
        issue_description:
          type: string
          maxLength: 10000
          nullable: true
        issue_image_url:
          type: string
          format: uri
          description: Screenshot of the fault. Uploaded separately via /documents/presign.
          nullable: true
        priority:
          $ref: '#/components/schemas/TicketPriority'
    SupportTicket:
      type: object
      required:
      - id
      - ticket_id
      - category
      - issue_summary
      - status
      - priority
      - created_at
      - updated_at
      properties:
        id:
          type: string
          format: uuid
        ticket_id:
          type: string
          description: |-
            Human reference, unique per category. `TIC-<epoch>` for
            customers, `DS<epoch>` for dealers — UNIQUE (category, ticket_id)
            so the two streams stay separable and a future `internal`
            category can reuse either prefix.
          example: TIC-316519
        category:
          type: string
          enum:
          - customer
          - dealer
          - internal
          description: Derived from the caller's role. **Not accepted from the body.**
        caller_id:
          type: string
          format: uuid
          description: Null if the caller's profile was later deleted; `caller_email` survives for the
            audit trail.
          nullable: true
        caller_email:
          type: string
          format: email
          description: Denormalised so the ticket stays readable after account deletion.
          nullable: true
        issue_summary:
          type: string
        issue_description:
          type: string
          nullable: true
        issue_image_url:
          type: string
          format: uri
          nullable: true
        status:
          $ref: '#/components/schemas/TicketStatus'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        notes:
          type: array
          description: Agent timeline. Append-only — an entry is never edited or removed.
          items:
            type: object
            properties:
              ts:
                type: string
                format: date-time
              text:
                type: string
              agent:
                type: string
                format: email
        rca:
          type: string
          description: Root-cause analysis, set when the ticket is resolved.
          nullable: true
        assigned_to:
          type: string
          format: uuid
          nullable: true
        sla_due_at:
          type: string
          format: date-time
          description: Derived from priority, so the queue can be sorted by urgency without recomputing
            it per row.
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
    SupportTicketList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SupportTicket'
        next_cursor:
          type: string
          nullable: true
    QueueTicketUpdateRequest:
      type: object
      description: |-
        At least one of `status`, `assigned_to`, `note` or `rca`. `note`
        **appends** to the timeline; it never replaces it.
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        assigned_to:
          type: string
          format: uuid
          nullable: true
        note:
          type: object
          description: Appended to `notes` when present.
          required:
          - text
          properties:
            text:
              type: string
              minLength: 1
              maxLength: 5000
          nullable: true
        rca:
          type: string
          maxLength: 10000
          nullable: true
    VerificationDecisionRequest:
      type: object
      required:
      - note
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 2000
          description: |-
            Required on both approve and deny. Appended to the decision trail
            and emailed to the applicant — a rejection with no stated reason
            is not acceptable to send.
    Verification:
      type: object
      required:
      - id
      - ticket_id
      - category
      - status
      - created_at
      - updated_at
      - email_sent
      properties:
        id:
          type: string
          format: uuid
        ticket_id:
          type: string
          example: DV1788793917
        category:
          type: string
          enum:
          - dealer
          - agent
          - customer
        profile_id:
          type: string
          format: uuid
          description: Subject of the verification. Null on an orphan row that needs backfilling or dropping.
          nullable: true
        applicant:
          allOf:
          - $ref: '#/components/schemas/PublicProfile'
          nullable: true
        status:
          $ref: '#/components/schemas/VerificationStatus'
        proof_docs:
          type: array
          description: S3 keys of uploaded proof. Empty array, never null.
          items:
            type: string
        notes:
          type: array
          description: Decision trail. Append-only.
          items:
            type: object
            properties:
              ts:
                type: string
                format: date-time
              text:
                type: string
              agent:
                type: string
                format: email
              decision:
                type: string
                enum:
                - approved
                - denied
                - null
                nullable: true
        decided_by:
          type: string
          format: uuid
          nullable: true
        decided_at:
          type: string
          format: date-time
          nullable: true
        email_sent:
          type: boolean
        account_activated:
          type: boolean
          description: True once the decision flipped `users.is_active` — only ever true on approve.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VerificationList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Verification'
        next_cursor:
          type: string
          nullable: true
    Member:
      allOf:
      - $ref: '#/components/schemas/Profile'
      - type: object
        properties:
          dealership_name:
            type: string
            nullable: true
          verification_id:
            type: string
            format: uuid
            description: The verification that gates this account, so an agent can act from the roster.
            nullable: true
          suspension_reason:
            type: string
            nullable: true
          last_login_at:
            type: string
            format: date-time
            nullable: true
    MemberList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        next_cursor:
          type: string
          nullable: true
...
  examples:
...
    TicketNew:
      summary: New customer ticket
      value:
        id: 812a926c-7600-4e48-9185-b337b5f02883
        ticket_id: TIC-316519
        category: customer
        caller_id: bdc31c9e-f6ed-40d0-a32f-706062f81d77
        caller_email: yuvi01@gmail.com
        issue_summary: The web page was more slow
        issue_description: Pages take 8+ seconds since yesterday evening.
        issue_image_url: null
        status: new
        priority: medium
        notes: []
        rca: null
        assigned_to: null
        sla_due_at: '2026-09-03 14:54:45+00:00'
        created_at: '2026-09-02 14:54:45+00:00'
        updated_at: '2026-09-02 14:54:45+00:00'
        resolved_at: null
        closed_at: null
    TicketInProgress:
      summary: Dealer ticket, assigned and in progress
      value:
        id: 649fa8a9-0c1f-45d2-b911-74d682a4f551
        ticket_id: DS9940692696
        category: dealer
        caller_id: 82d81b31-6a7c-4d8e-9f0a-1b2c3d4e5f60
        caller_email: rahul01@gmail.com
        issue_summary: Bidding is not working properly
        issue_description: Submitting a quote returns a 500.
        issue_image_url: https://cdn.drivedeal.example/tickets/bidding-error.png
        status: in-progress
        priority: medium
        notes:
        - ts: '2026-09-02 16:09:03.244000+00:00'
          text: Reproduced on staging — missing brand index on the quote join.
          agent: support02@gmail.com
        rca: null
        assigned_to: 4772bb1a-ee75-4f45-b8e6-53d70f6e42b1
        sla_due_at: '2026-09-03 21:03:24+00:00'
        created_at: '2026-09-02 21:03:24+00:00'
        updated_at: '2026-09-02 16:09:03+00:00'
        resolved_at: null
        closed_at: null
    VerificationPending:
      summary: A dealer waiting on a human
      value:
        id: 4cac1879-8502-42ba-8140-d321aba9b77
        ticket_id: DV1788793917
        category: dealer
        profile_id: c5d69c6d-f348-4bba-bfac-9f295b7984de
        applicant:
          id: c5d69c6d-f348-4bba-bfac-9f295b7984de
          role: dealer
          full_name: Carlos Mendez
          avatar_url: null
          dealership_name: Lone Star Motors
          branch_name: Austin
          seller_rating: null
          member_since: '2026-09-08 09:14:00+00:00'
        status: pending
        proof_docs:
        - dealer-licence.pdf
        - utility-bill.png
        notes: []
        decided_by: null
        decided_at: null
        email_sent: false
        account_activated: false
        created_at: '2026-09-08 09:14:22+00:00'
        updated_at: '2026-09-08 09:14:22+00:00'